Move proxies to `physical_query_plan/operator_proxy` subfolder (#748)
More files will be added to the `physical_query_plan` folder. To stay organized, it makes sense to put all proxies into a dedicated subfolder.
