+++ src/lib/compiler/physical_query_plan/operator_proxy/aggregate_operator_proxy.hpp
