+++ src/lib/compiler/physical_query_plan/operator_proxy/filter_operator_proxy.cpp
