+++ src/lib/compiler/physical_query_plan/operator_proxy/projection_operator_proxy.hpp
