+++ src/lib/compiler/physical_query_plan/operator_proxy/union_operator_proxy.cpp
