+++ src/lib/compiler/physical_query_plan/pqp_serialization.cpp
@@ -1,17 +1,17 @@
 #include "pqp_serialization.hpp"
 
-#include "aggregate_operator_proxy.hpp"
-#include "alias_operator_proxy.hpp"
-#include "exchange_operator_proxy.hpp"
-#include "export_operator_proxy.hpp"
-#include "filter_operator_proxy.hpp"
-#include "import_operator_proxy.hpp"
-#include "join_operator_proxy.hpp"
-#include "limit_operator_proxy.hpp"
-#include "partition_operator_proxy.hpp"
-#include "projection_operator_proxy.hpp"
-#include "sort_operator_proxy.hpp"
-#include "union_operator_proxy.hpp"
+#include "operator_proxy/aggregate_operator_proxy.hpp"
+#include "operator_proxy/alias_operator_proxy.hpp"
+#include "operator_proxy/exchange_operator_proxy.hpp"
+#include "operator_proxy/export_operator_proxy.hpp"
+#include "operator_proxy/filter_operator_proxy.hpp"
+#include "operator_proxy/import_operator_proxy.hpp"
+#include "operator_proxy/join_operator_proxy.hpp"
+#include "operator_proxy/limit_operator_proxy.hpp"
+#include "operator_proxy/partition_operator_proxy.hpp"
+#include "operator_proxy/projection_operator_proxy.hpp"
+#include "operator_proxy/sort_operator_proxy.hpp"
+#include "operator_proxy/union_operator_proxy.hpp"
 
 namespace skyrise {
