+++ src/test/lib/compiler/physical_query_plan/operator_proxy/import_operator_proxy_test.cpp
@@ -1,4 +1,4 @@
-#include "compiler/physical_query_plan/import_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/import_operator_proxy.hpp"
 
 #include <string>
 #include <vector>
