+++ src/test/lib/compiler/physical_query_plan/operator_proxy/operator_proxy_test.cpp
@@ -3,10 +3,10 @@
 
 #include <gtest/gtest.h>
 
-#include "compiler/physical_query_plan/alias_operator_proxy.hpp"
-#include "compiler/physical_query_plan/filter_operator_proxy.hpp"
-#include "compiler/physical_query_plan/import_operator_proxy.hpp"
-#include "compiler/physical_query_plan/union_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/alias_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/filter_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/import_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/union_operator_proxy.hpp"
 #include "expression/expression_functional.hpp"
 #include "types.hpp"
