+++ src/test/lib/compiler/physical_query_plan/operator_proxy/union_operator_proxy_test.cpp
@@ -1,11 +1,11 @@
-#include "compiler/physical_query_plan/union_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/union_operator_proxy.hpp"
 
 #include <string>
 #include <vector>
 
 #include <gtest/gtest.h>
 
-#include "compiler/physical_query_plan/import_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/import_operator_proxy.hpp"
 #include "types.hpp"
 
 namespace skyrise {
