+++ src/test/lib/compiler/physical_query_plan/pqp_utils_test.cpp
@@ -2,11 +2,11 @@
 
 #include <gtest/gtest.h>
 
-#include "compiler/physical_query_plan/export_operator_proxy.hpp"
-#include "compiler/physical_query_plan/filter_operator_proxy.hpp"
-#include "compiler/physical_query_plan/import_operator_proxy.hpp"
-#include "compiler/physical_query_plan/join_operator_proxy.hpp"
-#include "compiler/physical_query_plan/union_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/export_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/filter_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/import_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/join_operator_proxy.hpp"
+#include "compiler/physical_query_plan/operator_proxy/union_operator_proxy.hpp"
 #include "expression/expression_functional.hpp"
 #include "expression/pqp_column_expression.hpp"
